Fixes and event handler refactors
