+++ src/main/java/com/fouristhenumber/utilitiesinexcess/CommonProxy.java
@@ -1,6 +1,6 @@
 package com.fouristhenumber.utilitiesinexcess;
 
-import com.fouristhenumber.utilitiesinexcess.utils.SoundEventHandler;
+import com.fouristhenumber.utilitiesinexcess.utils.SoundVolumeChecks;
 
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
@@ -9,7 +9,7 @@ import cpw.mods.fml.common.event.FMLServerStartingEvent;
 
 public class CommonProxy {
 
-    public SoundEventHandler soundEventHandler;
+    public SoundVolumeChecks soundVolumeChecks;
 
     public void preInit(FMLPreInitializationEvent event) {
         // Config is handled in the early mixin loader (UIEMixinLoader)
@@ -19,7 +19,9 @@ public class CommonProxy {
         ModItems.init();
     }
 
-    public void init(FMLInitializationEvent event) {}
+    public void init(FMLInitializationEvent event) {
+        soundVolumeChecks = new SoundVolumeChecks();
+    }
 
     public void postInit(FMLPostInitializationEvent event) {}
