+++ src/main/java/com/fouristhenumber/utilitiesinexcess/common/tileentities/TileEntitySoundMuffler.java
@@ -1,24 +1,36 @@
 package com.fouristhenumber.utilitiesinexcess.common.tileentities;
 
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.audio.ISound;
+import net.minecraft.client.entity.EntityClientPlayerMP;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.network.NetworkManager;
 import net.minecraft.network.Packet;
 import net.minecraft.network.play.server.S35PacketUpdateTileEntity;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.client.event.sound.PlaySoundEvent17;
 
 import com.fouristhenumber.utilitiesinexcess.UtilitiesInExcess;
+import com.fouristhenumber.utilitiesinexcess.config.BlockConfig;
+import com.gtnewhorizon.gtnhlib.eventbus.EventBusSubscriber;
 
+import cpw.mods.fml.common.eventhandler.EventPriority;
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
+import cpw.mods.fml.relauncher.Side;
+
+@EventBusSubscriber(side = Side.CLIENT)
 public class TileEntitySoundMuffler extends TileEntity {
 
     boolean active;
 
     public void enableMuffler() {
-        UtilitiesInExcess.proxy.soundEventHandler
+        UtilitiesInExcess.proxy.soundVolumeChecks
             .putSoundMuffler(worldObj.provider.dimensionId, xCoord, yCoord, zCoord);
     }
 
     public void disableMuffler() {
-        UtilitiesInExcess.proxy.soundEventHandler
+        UtilitiesInExcess.proxy.soundVolumeChecks
             .removeSoundMuffler(worldObj.provider.dimensionId, xCoord, yCoord, zCoord);
     }
 
@@ -72,6 +84,86 @@ public class TileEntitySoundMuffler extends TileEntity {
         if (worldObj.isRemote) {
             if (active) enableMuffler();
             else disableMuffler();
+        }
+    }
+
+    @EventBusSubscriber.Condition
+    public static boolean shouldEventBusSubscribe() {
+        return BlockConfig.soundMuffler.enableSoundMuffler;
+    }
+
+    @SubscribeEvent(priority = EventPriority.HIGH)
+    public static void muffleSound(PlaySoundEvent17 event) {
+
+        ISound sound = event.sound;
+
+        EntityClientPlayerMP player = Minecraft.getMinecraft().thePlayer;
+        if (player == null) return;
+
+        double x = Math.floor(sound.getXPosF());
+        double y = Math.floor(sound.getYPosF());
+        double z = Math.floor(sound.getZPosF());
+
+        if (UtilitiesInExcess.proxy.soundVolumeChecks.isInSoundMufflerRange(player.dimension, x, y, z)) {
+            float reduction = BlockConfig.soundMuffler.soundMufflerReduction / 100f;
+            event.result = new MuffledSound(event.sound, reduction);
+            player.worldObj.spawnParticle("smoke", sound.getXPosF(), sound.getYPosF(), sound.getZPosF(), 0, 0.03, 0);
+        }
+    }
+
+    private static class MuffledSound implements ISound {
+
+        ISound base;
+        float reduction;
+
+        MuffledSound(ISound base, float reduction) {
+            this.base = base;
+            this.reduction = reduction;
+        }
+
+        @Override
+        public float getVolume() {
+            return base.getVolume() * reduction;
+        }
+
+        @Override
+        public ResourceLocation getPositionedSoundLocation() {
+            return base.getPositionedSoundLocation();
+        }
+
+        @Override
+        public boolean canRepeat() {
+            return base.canRepeat();
+        }
+
+        @Override
+        public int getRepeatDelay() {
+            return base.getRepeatDelay();
+        }
+
+        @Override
+        public float getPitch() {
+            return base.getPitch();
+        }
+
+        @Override
+        public float getXPosF() {
+            return base.getXPosF();
+        }
+
+        @Override
+        public float getYPosF() {
+            return base.getYPosF();
+        }
+
+        @Override
+        public float getZPosF() {
+            return base.getZPosF();
+        }
+
+        @Override
+        public AttenuationType getAttenuationType() {
+            return base.getAttenuationType();
         }
     }
 }
