+++ src/main/java/com/fouristhenumber/utilitiesinexcess/network/PacketRainMuffledSync.java
@@ -1,6 +1,6 @@
 package com.fouristhenumber.utilitiesinexcess.network;
 
-import static com.fouristhenumber.utilitiesinexcess.common.blocks.BlockRainMuffler.NBT_RAIN_MUFFLED;
+import static com.fouristhenumber.utilitiesinexcess.common.tileentities.TileEntityRainMuffler.NBT_RAIN_MUFFLED;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.player.EntityPlayer;
