+++ src/main/resources/assets/utilitiesinexcess/lang/en_US.lang
@@ -77,6 +77,8 @@ tile.cursed_earth.name=Cursed Earth
 tile.cursed_earth.desc=Spawns mobs aggressively
 
 utilitiesinexcess.config.block.cursed_earth=Cursed Earth Configuration
+utilitiesinexcess.config.block.sound_muffler=Sound Muffler Configuration
+utilitiesinexcess.config.block.rain_muffler=Rain Muffler Configuration
 
 tile.sound_muffler.name=Sound Muffler
 tile.sound_muffler.desc.1=Muffles sounds within %s blocks to %s%% volume
@@ -86,6 +88,6 @@ tile.rain_muffler.name=Rain Muffler
 tile.rain_muffler.desc.1=Silences weather sounds %s blocks
 tile.rain_muffler.desc.2=Supply with a redstone signal to turn off
 tile.rain_muffler.desc.3=Right click to toggle weather sounds globally
-tile.rain_muffler.global.enable=You dull your senses to the sound of the rain
-tile.rain_muffler.global.disable=You become aware of the sound of the rain
+tile.rain_muffler.chat.global_enable=You dull your senses to the sound of the rain
+tile.rain_muffler.chat.global_disable=You become aware of the sound of the rain
